r2021/sofa-dev : FIX: warnings with Windows (it should only remains warnings from WriteState.inl)

+++ modules/sofa/component/topology/MarchingCubeUtility.cpp
@@ -288,7 +288,7 @@ void  MarchingCubeUtility::applyConvolution(unsigned int CONVOLUTION_LENGTH, uns
 
 void MarchingCubeUtility::createConvolutionKernel(unsigned int CONVOLUTION_LENGTH, vector< float >  &convolutionKernel) const
 {
-    float c = 1.0 + CONVOLUTION_LENGTH/2;
+    int c = (1 + CONVOLUTION_LENGTH/2);
     float total = 0.0;
 
     convolutionKernel.resize(CONVOLUTION_LENGTH*CONVOLUTION_LENGTH*CONVOLUTION_LENGTH);
